config/requirements.txt
Django==4.0.6
django-redis==5.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
python-dotenv==0.20.0
Pillow==9.2.0
